[dev-tool] extract `sortPackageJson` helper
as it is useful outside of migrate-test command, which could be removed in the future.
